% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{daily_statistics}
\alias{daily_statistics}
\alias{dMaxAvg8h}
\alias{dMean}
\title{Daily statistics}
\usage{
dMaxAvg8h(data, value, time = "Time", point = "Point", req = 18)

dMean(data, value, time = "Time", point = "Point", req = 18)
}
\arguments{
\item{data}{data frame}

\item{value}{name of variable to be processed}

\item{time}{column containing time (\code{POSIXct})}

\item{point}{column containing point ID}

\item{req}{minimum valid required 8h averages to compute their daily max}
}
\value{
\code{dMaxAvg8h}: daily maxima of 8h rolling mean, for each point

\code{dMean}: daily averages, for each point
}
\description{
\code{dMaxAvg8h}: daily maximum of 8h rolling mean

\code{dMean}: daily average
}
