% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod.data}
\alias{mod.data}
\title{Forecasted concentrations}
\format{A data frame with 658944 rows and 4 variables:
\describe{
  \item{Time}{time in POSIXct format}
  \item{Var}{\code{"c_PM10"} (PM10), \code{"c_PM25"} (PM2.5), \code{"c_NO2"} (NO2) or \code{"c_O3"} (ozone)}
  \item{Point}{station code}
  \item{Value}{concentration (in \eqn{\mu g/m^3}{\mug/m3})}
}}
\source{
\url{http://www.arpa.fvg.it/}
}
\usage{
mod.data
}
\description{
A dataset containing hourly concentrations of PM10, PM2.5, NO2 and ozone forecasted at
background sites in the region of Friuli Venezia Giulia (Italy)
}
\keyword{datasets}
