% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs.no2}
\alias{obs.no2}
\title{Observed NO2 concentrations}
\format{A data frame with 202752 rows and 8 variables:
\describe{
  \item{Time}{time (POSIXct format)}
  \item{ID}{station code}
  \item{Station}{station name}
  \item{Value}{concentration (in \eqn{\mu g/m^3}{\mug/m3})}
  \item{ZoneType}{\code{"URB"} (urban), \code{"SBR"} (suburban) or \code{"RUR"} (rural)}
  \item{StationType}{always \code{"BKG"} (background)}
  \item{Lat}{latitude}
  \item{Lon}{longitude}
}}
\source{
\url{http://www.arpa.fvg.it/}
}
\usage{
obs.no2
}
\description{
A dataset containing hourly concentrations of NO2 measured at background sites in the
region of Friuli Venezia Giulia (Italy)
}
\keyword{datasets}
