% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parameters}
\alias{parameters}
\alias{params_U}
\alias{perc}
\alias{threshold}
\title{Pollutant-dependent parameters}
\usage{
params_U(pollutant = c("NO2", "O3", "PM10", "PM2.5"))

perc(pollutant = c("NO2", "O3", "PM10", "PM2.5"))

threshold(pollutant = c("NO2", "O3", "PM10", "PM2.5"))
}
\arguments{
\item{pollutant}{one of \code{"NO2"}, \code{"O3"}, \code{"PM10"}, \code{"PM2.5"}}
}
\value{
\code{params_U}: numeric list of parameters \code{U_RV95r}, \code{alpha}, \code{RV}, \code{Np},
        \code{Nnp}, as in tab.2, p.21 in Janssen et al., 2017

\code{perc}: numeric value in \code{[0,1]}, corresponding to the high percentile selected for the pollutant
(if possible, according to legislation): for hourly NO2 99.8\%, for the 8h daily maximum of ozone
92.9\%, for daily PM10 and PM2.5 90.4\%

\code{threshold}: numeric values (in \eqn{\mu g/m^3}{\mug/m3}) of the threshold used to
calculate exceedances, according to legislation:
200 for hourly NO2, 120 for the 8h daily maximum of ozone, 50 for daily PM10, \code{NA} otherwise
}
\description{
Parameters used to calculate the measurement uncertainty
}
\references{
Janssen et al., 2017. "Guidance Document on Modelling Quality Objectives and Benchmarking. Version 2.1"
}
