% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_uncertainty.R
\name{params_U}
\alias{params_U}
\title{Parameters for Measurements Uncertainty}
\usage{
params_U(pollutant = c("NO2", "O3", "PM10", "PM2.5"))
}
\arguments{
\item{pollutant}{one of \code{"NO2"}, \code{"O3"}, \code{"PM10"}, \code{"PM2.5"}}
}
\value{
numeric list of parameters \code{U_RV95r}, \code{alpha}, \code{RV}, \code{Np},
        \code{Nnp}, as in tab.2, p.21 in Janssen et al., 2017
}
\description{
Parameters used to calculate the measurement uncertainty
}
\references{
Janssen et al., 2017. "Guidance Document on Modelling Quality Objectives and Benchmarking. Version 2.1"
}
