% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_uncertainty.R
\name{RMS_U}
\alias{RMS_U}
\title{Root Mean Square Uncertainty}
\usage{
RMS_U(obs, pollutant = NULL, U_RV95r = NULL, alpha = NULL, RV = NULL)
}
\arguments{
\item{obs}{numeric vector of observed values}

\item{pollutant}{one of \code{"NO2"}, \code{"O3"}, \code{"PM10"}, \code{"PM2.5"}}

\item{U_RV95r}{parameter \eqn{U^{RV}_{95,r}}{U^{RV}_{95,r}}; optional if \code{pollutant} is given, otherwise compulsory}

\item{alpha}{parameter \eqn{\alpha}{\alpha}; optional if \code{pollutant} is given, otherwise compulsory}

\item{RV}{parameter \eqn{RV}{RV}; optional if \code{pollutant} is given, otherwise compulsory}
}
\value{
root mean square uncertainty \eqn{RMS_{U}}{RMS_U}, as in eq.11, p.20 in Janssen et al., 2017
}
\description{
Root Mean Square Uncertainty
}
