% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target_plot}
\alias{target_plot}
\alias{target_report}
\alias{target_plot}
\title{Target plot}
\usage{
target_report(data, obs, mod, point, pollutant = c("NO2", "O3", "PM10",
  "PM2.5"), beta = 2)

target_plot(t_rep)
}
\arguments{
\item{data}{a data frame including observations and model forecasts}

\item{obs}{name of the column with observed data}

\item{mod}{name of the column with forecasts}

\item{point}{name of the column with station ID}

\item{pollutant}{one of \code{"NO2"}, \code{"O3"}, \code{"PM10"}, \code{"PM2.5"}}

\item{beta}{parameter \eqn{\beta} (default is 2)}

\item{t_rep}{output of \code{target_report}}
}
\value{
\code{target_report} returns a list of 3:
\describe{
  \item{\code{quality_points}}{a data frame with 9 variables for each station \describe{
  \item{\code{Point}}{station ID}
  \item{\code{Obs.ave}}{annual mean of observations (paired with forecasts)}
  \item{\code{Mod.ave}}{annual mean of forecasts (paired with observations)}
  \item{\code{rmsu}}{\eqn{RMS_U} (see \link{RMS_U_obs})}
  \item{\code{crmse_norm}}{\eqn{CRMSE/(\beta*RMS_U)}{CRMSE/(\beta RMS_U)}}
  \item{\code{r}}{R (see \link{R})}
  \item{\code{crmse_ratio}}{\eqn{abs(NMSD)/sqrt(2*(1-R))}{\frac{|NMSD|}{\sqrt{2(1-R)}}} as in eq.34, p.30 in Janssen et al., 2017}
  \item{\code{bias_norm}}{\eqn{BIAS/(\beta*RMS_U)}{BIAS/(\beta RMS_U)}}
  \item{\code{mqi_ts}}{Model Quality Indicator \eqn{RMSE/(\beta*RMS_U)}{RMSE/(\beta RMS_U)} (see \link{MQI_ts_synth})}
  \item{\code{mqi_year}}{Model Quality Indicator for yearly averages (see \link{MQI_year})}
  \item{\code{n_valid}}{no. of valid data}
  }}
  \item{\code{quality_overall}}{a data frame with 3 overall indicators \describe{
  \item{\code{mqi_ts_p90}}{90th percentile of \code{mqi_ts}}
  \item{\code{mqi_year_p90}}{90th percentile of \code{mqi_year}}
  \item{\code{n_points}}{no. of valid stations}
  }}
  \item{\code{parameters}}{a list of 7 parameters \describe{
  \item{\code{U_RV95r}}{see \link{params_U}}
  \item{\code{alpha}}{see \link{params_U}}
  \item{\code{RV}}{see \link{params_U}}
  \item{\code{Np}}{see \link{params_U}}
  \item{\code{Nnp}}{see \link{params_U}}
  \item{\code{pollutant}}{same as in input, one of \code{"NO2"}, \code{"O3"}, \code{"PM10"}, \code{"PM2.5"}}
  \item{\code{beta}}{same as in input}
  }}
}

\code{target_plot} returns a target plot (object of class \code{ggplot} and \code{gg})
}
\description{
\code{target_report}: prepare input for target plot

\code{target_plot}: target plot
}
\examples{
# prepare dataset
require(dplyr)
Mod <- dMean(mod.data \%>\% filter(Var=="c_PM10"),
             value = "Value", time = "Time", point = "Point")
Obs <- obs.pm10 \%>\% mutate(Day=format(Time,"\%Y-\%m-\%d"),
                           Point=ID)
Dat <- inner_join(Mod, Obs, by=c("Point", "Day"), suffix = c(".mod", ".obs"))

# calculate indicators
t_rep <- target_report(Dat, obs = "Value.obs", mod = "Value.mod",
                       point = "Point", pollutant = "PM10")

# plot
target_plot(t_rep)
}
\references{
Janssen et al., 2017. "Guidance Document on Modelling Quality Objectives and Benchmarking. Version 2.1"
}
