% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators}
\alias{indicators}
\alias{RMSE}
\alias{R}
\alias{BIAS}
\alias{NMB}
\alias{NMSD}
\title{Core Set of Statistical Indicators}
\usage{
RMSE(obs, mod)

R(obs, mod)

BIAS(obs, mod)

NMB(obs, mod)

NMSD(obs, mod)
}
\arguments{
\item{obs}{numeric vector of observed values}

\item{mod}{numeric vector of modelled values}
}
\value{
\code{RMSE}: Root Mean Square Error as in eq.1, tab.1, p.18 in Janssen et al., 2017

\code{R}: Correlation coefficient as in eq.2, tab.1, p.18 in Janssen et al., 2017

\code{BIAS}: bias as in eq.3, tab.1, p.18 in Janssen et al., 2017

\code{NMB}: Normalised Mean Bias as in eq.3, tab.1, p.18 in Janssen et al., 2017

\code{NMSD}: Normalised Mean Standard Deviation as in eq.4, tab.1, p.18 in Janssen et al., 2017
}
\description{
\code{RMSE}: Root Mean Square Error

\code{R}: Correlation coefficient

\code{BIAS}: Bias

\code{NMB}: Normalised Mean Bias

\code{NMSD}: Normalised Mean Standard Deviation
}
\references{
Janssen et al., 2017. "Guidance Document on Modelling Quality Objectives and Benchmarking. Version 2.1"
}
