% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_uncertainty.R
\name{mod_uncertainty}
\alias{mod_uncertainty}
\alias{RMS_U_mod}
\alias{U_mod_year}
\title{Model Uncertainty}
\usage{
RMS_U_mod(obs, mod, pollutant, ...)

U_mod_year(obs, mod, pollutant, ...)
}
\arguments{
\item{obs}{numeric vector of observed values (yearly averages in \code{U_mod_year})}

\item{mod}{numeric vector of modelled values (yearly averages in \code{U_mod_year})}

\item{pollutant}{one of \code{"NO2"}, \code{"O3"}, \code{"PM10"}, \code{"PM2.5"}}

\item{...}{arguments to be passed to \code{U_obs_95} in \code{RMS_U_mod},
to \code{U_obs_95_year} in \code{U_mod_year}}
}
\value{
\code{RMS_U_mod}: root mean square uncertainty for model time series, as in eq.23, p.24 in Janssen et al., 2017

\code{MQI_ts}: uncertainty for model yearly averages, as in eq.24, p.24 in Janssen et al., 2017
}
\description{
\code{RMS_U_mod}: RMS uncertainty for model time series

\code{U_mod_year}: Uncertainty for model yearly averages
}
\references{
Janssen et al., 2017. "Guidance Document on Modelling Quality Objectives and Benchmarking. Version 2.1"
}
