% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_performance.R
\name{MPI}
\alias{MPI}
\alias{MPI_bias}
\alias{MPI_corr_time}
\alias{MPI_sdev_time}
\alias{MPI_corr_space}
\alias{MPI_sdev_space}
\title{Modelling Performance Indicators}
\usage{
MPI_bias(obs, mod, pollutant, beta = 2, ...)

MPI_corr_time(obs, mod, pollutant, beta = 2, ...)

MPI_sdev_time(obs, mod, pollutant, beta = 2, ...)

MPI_corr_space(obs, mod, pollutant, beta = 2, ...)

MPI_sdev_space(obs, mod, pollutant, beta = 2, ...)
}
\arguments{
\item{obs}{numeric vector of observed values (yearly averages in \code{MPI_*_space})}

\item{mod}{numeric vector of modelled values (yearly averages in \code{MPI_*_space})}

\item{pollutant}{one of \code{"NO2"}, \code{"O3"}, \code{"PM10"}, \code{"PM2.5"}}

\item{beta}{parameter \eqn{\beta} (default is 2)}

\item{...}{arguments to be passed to \code{RMS_U_obs} in \code{MPI_bias} and \code{MPI_*_time},
to \code{U_obs_95_year} in \code{MPI_*_space}}
}
\value{
\code{MPI_bias}: Modelling Performance Indicator for the bias, as in eq.26, tab.4, p.27 in Janssen et al., 2017

\code{MPI_corr_time}: temporal Modelling Performance Indicator for the correlation, as in eq.27, tab.4, p.27 in Janssen et al., 2017

\code{MPI_sdev_time}: temporal Modelling Performance Indicator for the standard deviation, as in eq.28, tab.4, p.27 in Janssen et al., 2017

\code{MPI_corr_space}: spatial Modelling Performance Indicator for the correlation, as in eq.29, tab.5, p.27 in Janssen et al., 2017

\code{MPI_sdev_space}: spatial Modelling Performance Indicator for the standard deviation, as in eq.30, tab.5, p.27 in Janssen et al., 2017
}
\description{
\code{MPI_bias}: MPI for bias

\code{MPI_corr_time}: temporal MPI for correlation

\code{MPI_sdev_time}: temporal MPI for standard deviation

\code{MPI_corr_space}: spatial MPI for correlation

\code{MPI_sdev_space}: spatial MPI for standard deviation
}
\references{
Janssen et al., 2017. "Guidance Document on Modelling Quality Objectives and Benchmarking. Version 2.1"
}
