% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quality.R
\name{MQI}
\alias{MQI}
\alias{MQI_ts}
\alias{MQI_ts_synth}
\alias{MQI_year}
\title{Modelling Quality Indicators}
\usage{
MQI_ts(obs, mod, pollutant, beta = 2, ...)

MQI_ts_synth(obs, mod, pollutant, beta = 2, ...)

MQI_year(obs, mod, pollutant, beta = 2, ...)
}
\arguments{
\item{obs}{numeric vector of observed values (yearly averages in \code{MQI_year})}

\item{mod}{numeric vector of modelled values (yearly averages in \code{MQI_year})}

\item{pollutant}{one of \code{"NO2"}, \code{"O3"}, \code{"PM10"}, \code{"PM2.5"}}

\item{beta}{parameter \eqn{\beta} (default is 2)}

\item{...}{arguments to be passed to \code{U_obs_95} in \code{MQI_ts} and \code{MQI_ts_synth},
to \code{U_obs_95_year} in \code{MQI_year}}
}
\value{
\code{MQI_ts}: Modelling Quality Indicator for a time series, as in eq.15, p.22 in Janssen et al., 2017

\code{MQI_ts_synth}: scalar synthetic Modelling Quality Indicator for a time series, as in eq.17, p.23 in Janssen et al., 2017

\code{MQI_year}: Modelling Quality Indicator for yearly average, as in eq.18, p.23 in Janssen et al., 2017
}
\description{
\code{MQI_ts}: MQI for a time series (vector)

\code{MQI_ts_synth}: MQI for a time series (synthetic scalar)

\code{MQI_year}: MQI for yearly average
}
\references{
Janssen et al., 2017. "Guidance Document on Modelling Quality Objectives and Benchmarking. Version 2.1"
}
